backend/requirements.txt
fastapi
sqlmodel
uvicorn
python-dotenv
python-jose[jwt]
passlib[bcrypt]
pytest
httpx